% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_r.R
\name{class}
\alias{class}
\title{COICOP Classification Codes}
\usage{
class()
}
\value{
It will return all the classified objects according to COICOP Classification
}
\description{
COICOP Classification Codes
}
