% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coicop_r.R
\name{cpi}
\alias{cpi}
\title{OECD Consumer Price Indice(s), CPI}
\usage{
cpi(country, class, start, end, period)
}
\arguments{
\item{country}{The code of the country}

\item{class}{The COICOP classification code}

\item{start}{Beginning year}

\item{end}{End year}

\item{period}{Select the period of data (Annualy, Quarterly, Monthly)}
}
\value{
The code will return a CPI dataset according to the components (country, class, start, end, period)
}
\description{
OECD Consumer Price Indice(s), CPI
}
\examples{
cpi(country = "DEU", class = "CPALTT01", start = 2000, end = 2010, period = "Q")
}
